基本情報の表示コンポーネントにスタイルを指定
